Move TokenColor class to parent dir
