+++ src/token/phpToken.ts
@@ -1,5 +1,5 @@
 import { tokenColorsDef } from "../color";
-import TokenColor from "./TokenColor";
+import TokenColor from "../TokenColor";
 
 export default function phpToken(token: TokenColor) {
   token.set(
