+++ src/tokenColors.ts
@@ -4,8 +4,8 @@ import makefileToken from "./token/makefileToken";
 import markupToken from "./token/markup";
 import phpToken from "./token/phpToken";
 import scssToken from "./token/scssToken";
-import TokenColor from "./token/TokenColor";
 import typescriptToken from "./token/typescript";
+import TokenColor from "./TokenColor";
 
 export default function getTokenColor() {
   const token = new TokenColor();
